f2fs: remove costly dirty_dir_inode operations
This patch removes list opeations in handling dirty dir inodes.
Previously, F2FS traverses whole the list of dirty dir inodes to check whether
there is an existing inode or not, resulting in heavy CPU overheads.

So this patch removes such the traverse operations by adding FI_DIRTY_DIR to
indicate the inode lies on the list or not.
Through this simple flag, we can remove redundant operations gracefully.

Change-Id: Ic82244ad5dd0eb4cfc7f71ae82008b43324c74c0
Signed-off-by: Jaegeuk Kim <jaegeuk.kim@samsung.com>